requirements.txt
python-telegram-bot==22.5
aiosqlite==0.20.0
